requirements.txt
nextcord==2.6.0
ffprobe-python